feat(web): convert /blog/[slug] and /leaderboard to hono/jsx
Two more page routes migrated to JSX, same verified pattern:
- blog/[slug]: post body via raw(mdToHtml(...)); the .md twin + 404 paths are
  unchanged string responses. (local raw slug renamed to avoid clashing with
  hono/html's raw import.)
- leaderboard: fragment helpers (RankRow/CategorySections/BuilderRows components),
  conditional headline/live-counter/ranked/by-builder sections, &-entities via raw().

Verified: blog.test.js (5/5) and leaderboard.test.js (3/3) pass, tsc --noEmit clean.
